GameOver analytics moved to GameOverScript
