+++ Assets/Scripts/GameOverScript.cs
@@ -1,5 +1,7 @@
 ﻿using UnityEngine;
 using System.Collections;
+using System.Collections.Generic;
+using UnityEngine.Cloud.Analytics;
 
 public class GameOverScript : MonoBehaviour
 {
@@ -12,6 +14,9 @@ public class GameOverScript : MonoBehaviour
 	private Rect button2Rect = new Rect(15,15,160,30);	
 
     public string herstartlevel;
+
+    // ANALYTICS
+    private bool playerRestart;
 
 	void Start()
 	{
@@ -51,6 +56,9 @@ public class GameOverScript : MonoBehaviour
 				"Opnieuw?"
 				)) 
 			{
+                playerRestart = true;
+                StartGameAnalytics();
+
 				GameOverActive = false;
 				gameoverMenu.gameObject.SetActive(false);
 				Time.timeScale = 1;
@@ -64,6 +72,9 @@ public class GameOverScript : MonoBehaviour
 				"Menu"
 				)) 
 			{
+                playerRestart = true;
+                StartGameAnalytics();
+
 				GameOverActive = false;
 				gameoverMenu.gameObject.SetActive(false);
 				Time.timeScale = 1;
@@ -71,4 +82,47 @@ public class GameOverScript : MonoBehaviour
 			}
 		}		
 	}
+
+    /**
+     * Sends the selected player and level to analytics
+     */
+    void StartGameAnalytics()
+    {
+        string customEventName = "PlayerDeath" + Application.loadedLevelName;
+
+        if (!GameControl.control.bossBattleStarted)
+        {
+            UnityAnalytics.CustomEvent(customEventName, new Dictionary<string, object>
+            {
+                { "runningTime", Time.timeSinceLevelLoad },
+                { "projectile1Shot", GameControl.control.projectile1Shot },
+                { "projectile2Shot", GameControl.control.projectile2Shot },
+                { "projectile3Shot", GameControl.control.projectile3Shot },
+                { "kidsFound", GameControl.control.kidsFound },
+                { "lettersFound", GameControl.control.lettersFound },
+                { "enemyDefeated", GameControl.control.enemiesDefeated },
+                { "respawns", GameControl.control.respawns },
+                { "timesPaused", GameControl.control.timesPaused },
+                { "playerRestart", playerRestart },
+            });
+        }
+        else
+        {
+            float bossBattleDuration = (Time.timeSinceLevelLoad - GameControl.control.bossBattleStartTime);
+
+            UnityAnalytics.CustomEvent(customEventName + "Boss", new Dictionary<string, object>
+            {
+                { "runningTime", Time.timeSinceLevelLoad },
+                { "projectile1Shot", GameControl.control.projectile1Shot },
+                { "projectile2Shot", GameControl.control.projectile2Shot },
+                { "projectile3Shot", GameControl.control.projectile3Shot },
+                { "kidsFound", GameControl.control.kidsFound },
+                { "bossBattleHealth", GameControl.control.bossDamageTaken },
+                { "bossBattleDuration", bossBattleDuration },
+                { "timesPaused", GameControl.control.timesPaused },
+                { "pauseDuration", GameControl.control.pauseDuration },
+                { "playerRestart", playerRestart },
+            });
+        }
+    }
 }
