+++ Assets/Scripts/Player.cs
@@ -1,8 +1,6 @@
 ﻿using UnityEngine;
 using UnityEngine.UI;
 using System.Collections;
-using System.Collections.Generic;
-using UnityEngine.Cloud.Analytics;
 
 public class Player : MonoBehaviour {
 
@@ -527,49 +525,5 @@ public class Player : MonoBehaviour {
 			}
 		}
 	}
-
-    /**
-     * Sends the selected player and level to analytics
-     * 
-     * Called by Respawn()
-     */
-    void StartGameAnalytics()
-    {
-        string customEventName = "PlayerDeath" + Application.loadedLevelName;
-
-        if (!GameControl.control.bossBattleStarted)
-        {
-            UnityAnalytics.CustomEvent(customEventName, new Dictionary<string, object>
-            {
-                { "runningTime", Time.timeSinceLevelLoad },
-                { "damageTaken", GameControl.control.damageTaken },
-                { "projectile1Shot", GameControl.control.projectile1Shot },
-                { "projectile2Shot", GameControl.control.projectile2Shot },
-                { "projectile3Shot", GameControl.control.projectile3Shot },
-                { "kidsFound", GameControl.control.kidsFound },
-                { "lettersFound", GameControl.control.lettersFound },
-                { "enemyDefeated", GameControl.control.enemiesDefeated },
-                { "respawns", GameControl.control.respawns },
-                { "timesPaused", GameControl.control.timesPaused },
-            });
-        }
-        else
-        {
-            float bossBattleDuration = (Time.timeSinceLevelLoad - GameControl.control.bossBattleStartTime);
-
-            UnityAnalytics.CustomEvent(customEventName + "Boss", new Dictionary<string, object>
-            {
-                { "runningTime", Time.timeSinceLevelLoad },
-                { "damageTaken", GameControl.control.damageTaken },
-                { "projectile1Shot", GameControl.control.projectile1Shot },
-                { "projectile2Shot", GameControl.control.projectile2Shot },
-                { "projectile3Shot", GameControl.control.projectile3Shot },
-                { "bossBattleHealth", GameControl.control.bossDamageTaken },
-                { "bossBattleDuration", bossBattleDuration },
-                { "timesPaused", GameControl.control.timesPaused },
-                { "pauseDuration", GameControl.control.pauseDuration },
-            });
-        }
-    }
 }
